Internal/master

+++ Packages/com.unity.render-pipelines.core/Runtime/Lighting/ProbeVolume/ProbeReferenceVolume.Streaming.cs
@@ -7,12 +7,22 @@ namespace UnityEngine.Rendering
         DynamicArray<CellInfo> m_TempCellToLoadList = new DynamicArray<CellInfo>();
         DynamicArray<CellInfo> m_TempCellToUnloadList = new DynamicArray<CellInfo>();
 
-        DynamicArray<BlendingCellInfo> m_LoadedBlendingCells = new();
-        DynamicArray<BlendingCellInfo> m_ToBeLoadedBlendingCells = new();
-        DynamicArray<BlendingCellInfo> m_TempBlendingCellToLoadList = new();
-        DynamicArray<BlendingCellInfo> m_TempBlendingCellToUnloadList = new();
+        DynamicArray<BlendingCellInfo> m_LoadedBlendingCells = new ();
+        DynamicArray<BlendingCellInfo> m_ToBeLoadedBlendingCells = new ();
+        DynamicArray<BlendingCellInfo> m_TempBlendingCellToLoadList = new ();
+        DynamicArray<BlendingCellInfo> m_TempBlendingCellToUnloadList = new ();
 
         Vector3 m_FrozenCameraPosition;
+        Vector3 m_FrozenCameraDirection;
+
+        const float kIndexFragmentationThreshold = 0.2f;
+        bool m_IndexDefragmentationInProgress;
+        ProbeBrickIndex m_DefragIndex;
+        ProbeGlobalIndirection m_DefragCellIndices;
+        DynamicArray<CellInfo> m_IndexDefragCells = new DynamicArray<CellInfo>();
+
+        internal float minStreamingScore;
+        internal float maxStreamingScore;
 
         bool m_HasRemainingCellsToBlend = false;
         internal void ScenarioBlendingChanged(bool scenarioChanged)
@@ -35,13 +45,16 @@ namespace UnityEngine.Rendering
             m_NumberOfCellsLoadedPerFrame = Mathf.Max(1, numberOfCells);
         }
 
-        void ComputeCellCameraDistance(Vector3 cameraPosition, DynamicArray<CellInfo> cells)
+        void ComputeStreamingScore(Vector3 cameraPosition, Vector3 cameraDirection, DynamicArray<CellInfo> cells)
         {
             for (int i = 0; i < cells.size; ++i)
             {
                 var cellInfo = cells[i];
-                // For now streaming score is only distance based.
+                var cellPosition = cellInfo.cell.position;
+                var cameraToCell = (cellPosition - cameraPosition).normalized;
                 cellInfo.streamingScore = Vector3.Distance(cameraPosition, cellInfo.cell.position);
+                // This should give more weight to cells in front of the camera.
+                cellInfo.streamingScore *= (2.0f - Vector3.Dot(cameraDirection, cameraToCell));
             }
         }
 
@@ -68,8 +81,7 @@ namespace UnityEngine.Rendering
             if (cellInfo.cell.shChunkCount <= shBudget && cellInfo.cell.indexChunkCount <= indexBudget)
             {
                 // This can still fail because of fragmentation.
-                // TODO: Handle defrag
-                if (LoadCell(cellInfo))
+                if (LoadCell(cellInfo, ignoreErrorLog: true))
                 {
                     loadedCells.Add(cellInfo);
 
@@ -98,6 +110,24 @@ namespace UnityEngine.Rendering
             return true;
         }
 
+        void ComputeMinMaxStreamingScore()
+        {
+            minStreamingScore = float.MaxValue;
+            maxStreamingScore = float.MinValue;
+
+            if (m_ToBeLoadedCells.size != 0)
+            {
+                minStreamingScore = Mathf.Min(minStreamingScore, m_ToBeLoadedCells[0].streamingScore);
+                maxStreamingScore = Mathf.Max(maxStreamingScore, m_ToBeLoadedCells[m_ToBeLoadedCells.size - 1].streamingScore);
+            }
+
+            if (m_LoadedCells.size != 0)
+            {
+                minStreamingScore = Mathf.Min(minStreamingScore, m_LoadedCells[0].streamingScore);
+                maxStreamingScore = Mathf.Max(maxStreamingScore, m_LoadedCells[m_LoadedCells.size - 1].streamingScore);
+            }
+        }
+
         /// <summary>
         /// Updates the cell streaming for a <see cref="Camera"/>
         /// </summary>
@@ -113,16 +143,19 @@ namespace UnityEngine.Rendering
                 if (!probeVolumeDebug.freezeStreaming)
                 {
                     m_FrozenCameraPosition = cameraPosition;
+                    m_FrozenCameraDirection = camera.transform.forward;
                 }
 
                 // Cell position in cell space is the top left corner. So we need to shift the camera position by half a cell to make things comparable.
                 var cameraPositionCellSpace = (m_FrozenCameraPosition - m_Transform.posWS) / MaxBrickSize() - Vector3.one * 0.5f;
 
-                ComputeCellCameraDistance(cameraPositionCellSpace, m_ToBeLoadedCells);
-                ComputeCellCameraDistance(cameraPositionCellSpace, m_LoadedCells);
+                ComputeStreamingScore(cameraPositionCellSpace, m_FrozenCameraDirection, m_ToBeLoadedCells);
+                ComputeStreamingScore(cameraPositionCellSpace, m_FrozenCameraDirection, m_LoadedCells);
 
                 m_ToBeLoadedCells.QuickSort();
                 m_LoadedCells.QuickSort();
+
+                ComputeMinMaxStreamingScore();
 
                 // This is only a rough budget estimate at first.
                 // It doesn't account for fragmentation.
@@ -132,47 +165,66 @@ namespace UnityEngine.Rendering
 
                 if (m_SupportStreaming)
                 {
-                    while (m_TempCellToLoadList.size < cellCountToLoad)
+                    if (m_IndexDefragmentationInProgress)
                     {
-                        // Enough memory, we can safely load the cell.
-                        var cellInfo = m_ToBeLoadedCells[m_TempCellToLoadList.size];
-                        if (!TryLoadCell(cellInfo, ref shChunkBudget, ref indexChunkBudget, m_TempCellToLoadList))
-                            break;
+                        UpdateIndexDefragmentation();
                     }
-
-                    // Budget reached. We need to figure out if we can safely unload other cells to make room.
-                    if (m_TempCellToLoadList.size != cellCountToLoad)
+                    else
                     {
-                        int pendingUnloadCount = 0;
+                        bool needComputeFragmentation = false;
+
                         while (m_TempCellToLoadList.size < cellCountToLoad)
                         {
-                            if (m_LoadedCells.size - pendingUnloadCount == 0)
-                                break;
-
-                            var furthestLoadedCell = m_LoadedCells[m_LoadedCells.size - pendingUnloadCount - 1];
-                            var closestUnloadedCell = m_ToBeLoadedCells[m_TempCellToLoadList.size];
-
-                            // The most distant loaded cell is further than the closest unloaded cell, we can unload it.
-                            if (furthestLoadedCell.streamingScore > closestUnloadedCell.streamingScore)
-                            {
-                                pendingUnloadCount++;
-                                UnloadCell(furthestLoadedCell);
-                                shChunkBudget += furthestLoadedCell.cell.shChunkCount;
-                                indexChunkBudget += furthestLoadedCell.cell.indexChunkCount;
-
-                                m_TempCellToUnloadList.Add(furthestLoadedCell);
-
-                                TryLoadCell(closestUnloadedCell, ref shChunkBudget, ref indexChunkBudget, m_TempCellToLoadList);
-                            }
-                            else // We are in a "stable" state, all the closest cells are loaded within the budget.
+                            // Enough memory, we can safely load the cell.
+                            var cellInfo = m_ToBeLoadedCells[m_TempCellToLoadList.size];
+                            if (!TryLoadCell(cellInfo, ref shChunkBudget, ref indexChunkBudget, m_TempCellToLoadList))
                                 break;
                         }
 
-                        if (pendingUnloadCount > 0)
+                        // Budget reached. We need to figure out if we can safely unload other cells to make room.
+                        // If defrag was triggered by TryLoadCell we should not try to load further cells either.
+                        if (m_TempCellToLoadList.size != cellCountToLoad && !m_IndexDefragmentationInProgress)
                         {
-                            m_LoadedCells.RemoveRange(m_LoadedCells.size - pendingUnloadCount, pendingUnloadCount);
-                            RecomputeMinMaxLoadedCellPos();
+                            int pendingUnloadCount = 0;
+                            while (m_TempCellToLoadList.size < cellCountToLoad)
+                            {
+                                if (m_LoadedCells.size - pendingUnloadCount == 0)
+                                    break;
+
+                                var furthestLoadedCell = m_LoadedCells[m_LoadedCells.size - pendingUnloadCount - 1];
+                                var closestUnloadedCell = m_ToBeLoadedCells[m_TempCellToLoadList.size];
+
+                                // The most distant loaded cell is further than the closest unloaded cell, we can unload it.
+                                if (furthestLoadedCell.streamingScore > closestUnloadedCell.streamingScore)
+                                {
+                                    pendingUnloadCount++;
+                                    UnloadCell(furthestLoadedCell);
+                                    shChunkBudget += furthestLoadedCell.cell.shChunkCount;
+                                    indexChunkBudget += furthestLoadedCell.cell.indexChunkCount;
+
+                                    m_TempCellToUnloadList.Add(furthestLoadedCell);
+
+                                    if (!TryLoadCell(closestUnloadedCell, ref shChunkBudget, ref indexChunkBudget, m_TempCellToLoadList))
+                                        break; // Alloc failed because of fragmentation, stop trying to load cells.
+
+                                    needComputeFragmentation = true;
+                                }
+                                else // We are in a "stable" state, all the closest cells are loaded within the budget.
+                                    break;
+                            }
+
+                            if (pendingUnloadCount > 0)
+                            {
+                                m_LoadedCells.RemoveRange(m_LoadedCells.size - pendingUnloadCount, pendingUnloadCount);
+                                RecomputeMinMaxLoadedCellPos();
+                            }
                         }
+
+                        if (needComputeFragmentation)
+                            m_Index.ComputeFragmentationRate();
+
+                        if (m_Index.fragmentationRate >= kIndexFragmentationThreshold)
+                            StartIndexDefragmentation();
                     }
                 }
                 else
@@ -186,7 +238,9 @@ namespace UnityEngine.Rendering
 
                 // Remove the cells we successfully loaded.
                 m_ToBeLoadedCells.RemoveRange(0, m_TempCellToLoadList.size);
+                // Move the successfully loaded cells to the "loaded cells" list.
                 m_LoadedCells.AddRange(m_TempCellToLoadList);
+                // Remove the successfully loaded cells from the list of cells to be loaded.
                 m_ToBeLoadedCells.AddRange(m_TempCellToUnloadList);
                 m_TempCellToLoadList.Clear();
                 m_TempCellToUnloadList.Clear();
@@ -312,6 +366,66 @@ namespace UnityEngine.Rendering
 
             if (m_ToBeLoadedBlendingCells.size == 0)
                 m_HasRemainingCellsToBlend = false;
+        }
+
+        static int DefragComparer(CellInfo a, CellInfo b)
+        {
+            if (a.updateInfo.GetNumberOfChunks() > b.updateInfo.GetNumberOfChunks())
+                return 1;
+            else if (a.updateInfo.GetNumberOfChunks() < b.updateInfo.GetNumberOfChunks())
+                return -1;
+            else return 0;
+        }
+
+        void StartIndexDefragmentation()
+        {
+            m_IndexDefragmentationInProgress = true;
+
+            // Prepare the list of cells.
+            // We want to relocate cells with more indices first.
+            m_IndexDefragCells.Clear();
+            m_IndexDefragCells.AddRange(m_LoadedCells);
+            m_IndexDefragCells.QuickSort(DefragComparer);
+
+            m_DefragIndex.Clear();
+        }
+
+        void UpdateIndexDefragmentation()
+        {
+            using (new ProfilingScope(null, ProfilingSampler.Get(CoreProfileId.APVIndexDefragUpdate)))
+            {
+                int numberOfCellsToProcess = Mathf.Min(m_IndexDefragCells.size, m_NumberOfCellsLoadedPerFrame);
+                for (int i = 0; i < numberOfCellsToProcess; ++i)
+                {
+                    var cellInfo = m_IndexDefragCells[m_IndexDefragCells.size - i - 1];
+
+                    int indirectionBufferEntries = cellInfo.updateInfo.entriesInfo.Length;
+                    bool canAllocateCell = m_DefragIndex.FindSlotsForEntries(ref cellInfo.updateInfo.entriesInfo);
+                    bool successfulReserve = m_DefragIndex.ReserveChunks(cellInfo.updateInfo.entriesInfo, false);
+
+                    // Update index and indirection
+                    m_DefragIndex.AddBricks(cellInfo.cell, cellInfo.cell.bricks, cellInfo.chunkList, ProbeBrickPool.GetChunkSizeInBrickCount(), m_Pool.GetPoolWidth(), m_Pool.GetPoolHeight(), cellInfo.updateInfo);
+                    m_DefragCellIndices.UpdateCell(cellInfo.flatIndicesInGlobalIndirection, cellInfo.updateInfo);
+                }
+
+                // Remove processed cells from the list.
+                m_IndexDefragCells.Resize(m_IndexDefragCells.size - numberOfCellsToProcess);
+
+                if (m_IndexDefragCells.size == 0)
+                {
+                    // Swap index buffers
+                    var oldDefragIndex = m_DefragIndex;
+                    m_DefragIndex = m_Index;
+                    m_Index = oldDefragIndex;
+
+                    var oldDefragCellIndices = m_DefragCellIndices;
+                    m_DefragCellIndices = m_CellIndices;
+                    m_CellIndices = oldDefragCellIndices;
+
+                    // Resume streaming
+                    m_IndexDefragmentationInProgress = false;
+                }
+            }
         }
     }
 }
